jeff_agent/pyproject.toml
[project]
name = "jeff-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.24",
    "google-adk>=1.26.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "python-dotenv>=1.2.1",
]
